+++ packages/react-core/src/components/Modal/ModalBox.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { css } from '@patternfly/react-styles';
 import styles from '@patternfly/react-styles/css/components/ModalBox/modal-box';
+import topSpacer from '@patternfly/react-tokens/dist/js/c_modal_box_m_align_top_spacer';
 
 export interface ModalBoxProps extends React.HTMLProps<HTMLDivElement> {
   /** Content rendered inside the ModalBox. */
@@ -9,6 +10,10 @@ export interface ModalBoxProps extends React.HTMLProps<HTMLDivElement> {
   className?: string;
   /** Variant of the modal */
   variant?: 'small' | 'medium' | 'large' | 'default';
+  /** Alternate position of the modal */
+  position?: 'top';
+  /** Offset from alternate position. Can be any valid CSS length/percentage */
+  positionOffset?: string;
   /** Id to use for Modal Box label */
   'aria-labelledby'?: string;
   /** Accessible descriptor of modal */
@@ -21,27 +26,38 @@ export const ModalBox: React.FunctionComponent<ModalBoxProps> = ({
   children,
   className = '',
   variant = 'default',
+  position,
+  positionOffset,
   'aria-labelledby': ariaLabelledby,
   'aria-label': ariaLabel = '',
   'aria-describedby': ariaDescribedby,
+  style,
   ...props
-}: ModalBoxProps) => (
-  <div
-    {...props}
-    role="dialog"
-    aria-label={ariaLabel || null}
-    aria-labelledby={ariaLabelledby || null}
-    aria-describedby={ariaDescribedby}
-    aria-modal="true"
-    className={css(
-      styles.modalBox,
-      className,
-      variant === 'large' && styles.modifiers.lg,
-      variant === 'small' && styles.modifiers.sm,
-      variant === 'medium' && styles.modifiers.md
-    )}
-  >
-    {children}
-  </div>
-);
+}: ModalBoxProps) => {
+  if (positionOffset) {
+    style = style || {};
+    (style as any)[topSpacer.name] = positionOffset;
+  }
+  return (
+    <div
+      {...props}
+      role="dialog"
+      aria-label={ariaLabel || null}
+      aria-labelledby={ariaLabelledby || null}
+      aria-describedby={ariaDescribedby}
+      aria-modal="true"
+      className={css(
+        styles.modalBox,
+        className,
+        position === 'top' && styles.modifiers.alignTop,
+        variant === 'large' && styles.modifiers.lg,
+        variant === 'small' && styles.modifiers.sm,
+        variant === 'medium' && styles.modifiers.md
+      )}
+      style={style}
+    >
+      {children}
+    </div>
+  );
+};
 ModalBox.displayName = 'ModalBox';