+++ packages/react-icons/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@4.7.9...@patternfly/react-icons@4.7.10) (2020-10-02)
+
+**Note:** Version bump only for package @patternfly/react-icons
+
+
+
+
+
+## [4.7.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@4.7.8...@patternfly/react-icons@4.7.9) (2020-10-01)
+
+**Note:** Version bump only for package @patternfly/react-icons
+
+
+
+
+
+## [4.7.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@4.7.7...@patternfly/react-icons@4.7.8) (2020-09-28)
+
+**Note:** Version bump only for package @patternfly/react-icons
+
+
+
+
+
 ## [4.7.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@4.7.6...@patternfly/react-icons@4.7.7) (2020-09-23)
 
 **Note:** Version bump only for package @patternfly/react-icons