+++ packages/react-integration/demo-app-ts/CHANGELOG.md
@@ -3,6 +3,131 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.49.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.48.0...demo-app-ts@4.49.0) (2020-10-02)
+
+
+### Features
+
+* **Label:** Added support for React router ([#4934](https://github.com/patternfly/patternfly-react/issues/4934)) ([7f3a9ad](https://github.com/patternfly/patternfly-react/commit/7f3a9add2650cef8c8884ffdbaeb6663187775b0))
+
+
+
+
+
+# [4.48.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.47.3...demo-app-ts@4.48.0) (2020-10-02)
+
+
+### Features
+
+* **DataList:** add support for text modifiers ([#4846](https://github.com/patternfly/patternfly-react/issues/4846)) ([4f4271b](https://github.com/patternfly/patternfly-react/commit/4f4271b2c20a869504339173c28300ed65cd7330)), closes [#4822](https://github.com/patternfly/patternfly-react/issues/4822)
+
+
+
+
+
+## [4.47.3](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.47.2...demo-app-ts@4.47.3) (2020-10-02)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [4.47.2](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.47.1...demo-app-ts@4.47.2) (2020-10-02)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [4.47.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.47.0...demo-app-ts@4.47.1) (2020-10-02)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+# [4.47.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.46.3...demo-app-ts@4.47.0) (2020-10-01)
+
+
+### Features
+
+* **Modal:** add top align to modal ([#4871](https://github.com/patternfly/patternfly-react/issues/4871)) ([15da8a0](https://github.com/patternfly/patternfly-react/commit/15da8a08867899b20dc774aafcc6bfe458ac9ae0))
+
+
+
+
+
+## [4.46.3](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.46.2...demo-app-ts@4.46.3) (2020-10-01)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [4.46.2](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.46.1...demo-app-ts@4.46.2) (2020-10-01)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [4.46.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.46.0...demo-app-ts@4.46.1) (2020-10-01)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+# [4.46.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.45.4...demo-app-ts@4.46.0) (2020-09-30)
+
+
+### Features
+
+* **Progress:** added truncation ([#4842](https://github.com/patternfly/patternfly-react/issues/4842)) ([3b5f1ba](https://github.com/patternfly/patternfly-react/commit/3b5f1baba4a452abb45f48fb0968be703999dfd4))
+* **Select:** Added support for Divider in Select ([#4772](https://github.com/patternfly/patternfly-react/issues/4772)) ([4071f6f](https://github.com/patternfly/patternfly-react/commit/4071f6fd70d55356f16eebfac9bac3c24af1ed8d))
+
+
+
+
+
+## [4.45.4](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.45.3...demo-app-ts@4.45.4) (2020-09-30)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [4.45.3](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.45.2...demo-app-ts@4.45.3) (2020-09-28)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [4.45.2](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.45.1...demo-app-ts@4.45.2) (2020-09-28)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [4.45.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.45.0...demo-app-ts@4.45.1) (2020-09-28)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
 # [4.45.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@4.44.0...demo-app-ts@4.45.0) (2020-09-25)