+++ packages/react-table/CHANGELOG.md
@@ -3,6 +3,129 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.18.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.18.9...@patternfly/react-table@4.18.10) (2020-10-02)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.18.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.18.8...@patternfly/react-table@4.18.9) (2020-10-02)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.18.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.18.7...@patternfly/react-table@4.18.8) (2020-10-02)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.18.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.18.6...@patternfly/react-table@4.18.7) (2020-10-02)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.18.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.18.5...@patternfly/react-table@4.18.6) (2020-10-02)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.18.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.18.4...@patternfly/react-table@4.18.5) (2020-10-02)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.18.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.18.3...@patternfly/react-table@4.18.4) (2020-10-01)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.18.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.18.2...@patternfly/react-table@4.18.3) (2020-10-01)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.18.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.18.1...@patternfly/react-table@4.18.2) (2020-10-01)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.18.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.18.0...@patternfly/react-table@4.18.1) (2020-10-01)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+# [4.18.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.17.11...@patternfly/react-table@4.18.0) (2020-09-30)
+
+
+### Features
+
+* **Progress:** added truncation ([#4842](https://github.com/patternfly/patternfly-react/issues/4842)) ([3b5f1ba](https://github.com/patternfly/patternfly-react/commit/3b5f1baba4a452abb45f48fb0968be703999dfd4))
+
+
+
+
+
+## [4.17.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.17.10...@patternfly/react-table@4.17.11) (2020-09-30)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.17.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.17.9...@patternfly/react-table@4.17.10) (2020-09-28)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.17.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.17.8...@patternfly/react-table@4.17.9) (2020-09-28)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [4.17.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.17.7...@patternfly/react-table@4.17.8) (2020-09-28)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
 ## [4.17.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@4.17.6...@patternfly/react-table@4.17.7) (2020-09-25)
 
 **Note:** Version bump only for package @patternfly/react-table